feat: pruebas automatizadas para endpoints de productos
